requirements.txt
asgiref==3.8.1
astroid==3.3.5
black==24.10.0
click==8.1.7
dill==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
phonenumbers==8.13.51
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.1
pylint==3.3.2
sqlparse==0.5.2
tomlkit==0.13.2
whitenoise==6.8.2
